Faster quantile()

A faster version of quantile(), copied from this gist: https://gist.github.com/sikli/f1775feb9736073cefee97ec81f6b193

quantile_speed(x, probs = c(0.1, 0.9), na.rm = FALSE)

Arguments

x

numeric vector whose sample quantiles are wanted, or an object of a class for which a method has been defined (see also 'details'). NA and NaN values are not allowed in numeric vectors unless na.rm = TRUE.

probs

numeric vector of probabilities with values in [0,1]. (Values up to '2e-14' outside that range are accepted and moved to the nearby endpoint.)

na.rm

logical; if TRUE, any NA and NaN's are removed from x before the quantiles are computed.

Examples

quantile(x <- rnorm(1001)) # Extremes & Quartiles by default
#>          0%         25%         50%         75%        100% 
#> -2.88703458 -0.70525584 -0.03677616  0.65852553  3.46688299 
quantile(x,  probs = c(0.1, 0.5, 1, 2, 5, 10, 50, NA)/100)
#>        0.1%        0.5%          1%          2%          5%         10% 
#> -2.71290584 -2.49062423 -2.26980598 -2.01648724 -1.60353628 -1.31575151 
#>         50%             
#> -0.03677616          NA