Download a zipped file from the web and return the temporary paths to each file

This function downloads and locally stores a zipped file from a url and then returns a list of the paths to each file

read_zip(url)

Arguments

url

A url to a zipped file

Examples

# Download and view fuel economy data for all U.S. vehicles
# from 1980 to present (https://www.fueleconomy.gov/feg/download.shtml):
url <- 'https://www.fueleconomy.gov/feg/epadata/vehicles.csv.zip'
filePaths = read_zip(url)
print(filePaths)
#> [1] "/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmplZna9C/vehicles.csv"

# View first data file:
data <- read.csv(filePaths[1], header = TRUE)
head(data)
#>   barrels08 barrelsA08 charge120 charge240 city08 city08U cityA08 cityA08U
#> 1  14.16714          0         0         0     19       0       0        0
#> 2  27.04636          0         0         0      9       0       0        0
#> 3  11.01889          0         0         0     23       0       0        0
#> 4  27.04636          0         0         0     10       0       0        0
#> 5  15.65842          0         0         0     17       0       0        0
#> 6  13.52318          0         0         0     21       0       0        0
#>   cityCD cityE cityUF co2 co2A co2TailpipeAGpm co2TailpipeGpm comb08 comb08U
#> 1      0     0      0  -1   -1               0       423.1905     21       0
#> 2      0     0      0  -1   -1               0       807.9091     11       0
#> 3      0     0      0  -1   -1               0       329.1481     27       0
#> 4      0     0      0  -1   -1               0       807.9091     11       0
#> 5      0     0      0  -1   -1               0       467.7368     19       0
#> 6      0     0      0  -1   -1               0       403.9545     22       0
#>   combA08 combA08U combE combinedCD combinedUF cylinders displ
#> 1       0        0     0          0          0         4   2.0
#> 2       0        0     0          0          0        12   4.9
#> 3       0        0     0          0          0         4   2.2
#> 4       0        0     0          0          0         8   5.2
#> 5       0        0     0          0          0         4   2.2
#> 6       0        0     0          0          0         4   1.8
#>                        drive engId   eng_dscr feScore fuelCost08 fuelCostA08
#> 1           Rear-Wheel Drive  9011      (FFS)      -1       2400           0
#> 2           Rear-Wheel Drive 22020  (GUZZLER)      -1       4600           0
#> 3          Front-Wheel Drive  2100      (FFS)      -1       1900           0
#> 4           Rear-Wheel Drive  2850                 -1       4600           0
#> 5 4-Wheel or All-Wheel Drive 66031 (FFS,TRBO)      -1       3300           0
#> 6          Front-Wheel Drive 66020      (FFS)      -1       2300           0
#>   fuelType        fuelType1 ghgScore ghgScoreA highway08 highway08U highwayA08
#> 1  Regular Regular Gasoline       -1        -1        25          0          0
#> 2  Regular Regular Gasoline       -1        -1        14          0          0
#> 3  Regular Regular Gasoline       -1        -1        33          0          0
#> 4  Regular Regular Gasoline       -1        -1        12          0          0
#> 5  Premium Premium Gasoline       -1        -1        23          0          0
#> 6  Regular Regular Gasoline       -1        -1        24          0          0
#>   highwayA08U highwayCD highwayE highwayUF hlv hpv    id lv2 lv4       make
#> 1           0         0        0         0   0   0     1   0   0 Alfa Romeo
#> 2           0         0        0         0   0   0    10   0   0    Ferrari
#> 3           0         0        0         0  19  77   100   0   0      Dodge
#> 4           0         0        0         0   0   0  1000   0   0      Dodge
#> 5           0         0        0         0   0   0 10000   0  14     Subaru
#> 6           0         0        0         0   0   0 10001   0  15     Subaru
#>                 model mpgData phevBlended pv2 pv4 range rangeCity rangeCityA
#> 1  Spider Veloce 2000       Y       false   0   0     0         0          0
#> 2          Testarossa       N       false   0   0     0         0          0
#> 3             Charger       Y       false   0   0     0         0          0
#> 4 B150/B250 Wagon 2WD       N       false   0   0     0         0          0
#> 5    Legacy AWD Turbo       N       false   0  90     0         0          0
#> 6              Loyale       N       false   0  88     0         0          0
#>   rangeHwy rangeHwyA           trany   UCity UCityA UHighway UHighwayA
#> 1        0         0    Manual 5-spd 23.3333      0  35.0000         0
#> 2        0         0    Manual 5-spd 11.0000      0  19.0000         0
#> 3        0         0    Manual 5-spd 29.0000      0  47.0000         0
#> 4        0         0 Automatic 3-spd 12.2222      0  16.6667         0
#> 5        0         0    Manual 5-spd 21.0000      0  32.0000         0
#> 6        0         0 Automatic 3-spd 27.0000      0  33.0000         0
#>            VClass year youSaveSpend guzzler trans_dscr tCharger sCharger
#> 1     Two Seaters 1985        -3000                          NA         
#> 2     Two Seaters 1985       -14000       T                  NA         
#> 3 Subcompact Cars 1985         -500                SIL       NA         
#> 4            Vans 1985       -14000                          NA         
#> 5    Compact Cars 1993        -7500                        TRUE         
#> 6    Compact Cars 1993        -2500                          NA         
#>   atvType fuelType2 rangeA evMotor mfrCode c240Dscr charge240b c240bDscr
#> 1                                                            0          
#> 2                                                            0          
#> 3                                                            0          
#> 4                                                            0          
#> 5                                                            0          
#> 6                                                            0          
#>                      createdOn                   modifiedOn startStop phevCity
#> 1 Tue Jan 01 00:00:00 EST 2013 Tue Jan 01 00:00:00 EST 2013                  0
#> 2 Tue Jan 01 00:00:00 EST 2013 Tue Jan 01 00:00:00 EST 2013                  0
#> 3 Tue Jan 01 00:00:00 EST 2013 Tue Jan 01 00:00:00 EST 2013                  0
#> 4 Tue Jan 01 00:00:00 EST 2013 Tue Jan 01 00:00:00 EST 2013                  0
#> 5 Tue Jan 01 00:00:00 EST 2013 Tue Jan 01 00:00:00 EST 2013                  0
#> 6 Tue Jan 01 00:00:00 EST 2013 Tue Jan 01 00:00:00 EST 2013                  0
#>   phevHwy phevComb
#> 1       0        0
#> 2       0        0
#> 3       0        0
#> 4       0        0
#> 5       0        0
#> 6       0        0